Add files via upload
The dataset deals with the small-scale units in the undivided state of Andhra Pradesh. 
Carry out the following analysis. 

1.	Calculate a 95 percent confidence interval for the “Gross output – Year 3 (Rs)”
2.	Define two different measures that you consider most appropriate for measuring the performance of the units.  This definition is up to you.  These can be the variables that are already in the data or new variables defined based on the existing variables.  For example, you can define a metric “Gross output per employee” by dividing the gross output of the year by the total number of employees. Please explain in one paragraph why you have selected these two measures and why you think they are most appropriate. Remaining analysis is to be carried out based on these definitions.
3.	Calculate 99% confidence interval for the population mean for each of the two metrics defined by you.  Interpret these confidence intervals in terms of their relevance to the management.
4.	
a.	What is the probability that a firm selected at random is a SSSBE unit?
b.	What is the probability that a firm selected at random is GOOD in performance? (Calculate the average of the first performance measure that you had defined in question 2 above.  If the firm’s performance is above this average, it considered good. If it is below average, it is considered Bad)
c.	What is the probability that a firm selected is a SSSBE Unit and ALSO GOOD in performance? 
d.	What can you say about the performance of the SSSBE units in terms of GOOD or BAD based on the probabilities calculated above?
5.	Test the null hypothesis that the population average of the variable “Value of Exports for Year 3” = 87,300.  Carry out a one sided test.  Clearly state your null and alternate hypotheses.
6.	There is a feeling within the Central Government Department for promotion of small scale units, be it SSSBE or SSI, that if the population proportion is less than 25%, there is a need for providing special incentives.  Based on your sample, would you recommend these special incentives for SSSBE or SSI or both?
7.	Some male chauvinists like to think that a larger proportion of SSSBEs are managed by men as compared to women.  Do you agree with this contention?  Explain your answer with appropriate statistical evidence.
8.	Comment on the distribution of the two metrics (variables) created by you.  Can you conclude that they follow normal distribution?  Explain the rationale for your conclusion.